refactor(frontend): tighten shared package exports
