+++ apps/desktop/src/components/workspace/shell/sidebar/RepoGroup.tsx
@@ -17,7 +17,7 @@ import {
 import { useRepoGroupNativeContextMenu } from "@/hooks/workspaces/ui/use-repo-group-native-context-menu";
 import { useNewWorkspaceCommandScopeStore } from "@/stores/workspaces/new-workspace-command-scope-store";
 import { SidebarActionButton } from "@proliferate/ui/layout/SidebarActionButton";
-import { ProductSidebarRepoGroupHeader } from "@proliferate/product-ui/sidebar/ProductSidebar";
+import { ProductSidebarRepoGroupHeader } from "@proliferate/product-ui/sidebar/ProductSidebarRepositories";
 
 interface RepoGroupProps {
   name: string;
