+++ apps/desktop/src/components/workspace/shell/sidebar/SidebarRepositoriesHeader.tsx
@@ -9,7 +9,7 @@ import { PopoverButton } from "@proliferate/ui/primitives/PopoverButton";
 import { PopoverMenuItem } from "@proliferate/ui/primitives/PopoverMenuItem";
 import { SidebarActionButton } from "@proliferate/ui/layout/SidebarActionButton";
 import { SidebarWorkspaceVariantIcon } from "@/components/workspace/shell/sidebar/SidebarWorkspaceVariantIcon";
-import { ProductSidebarSectionHeader } from "@proliferate/product-ui/sidebar/ProductSidebar";
+import { ProductSidebarSectionHeader } from "@proliferate/product-ui/sidebar/ProductSidebarLayout";
 import type { SidebarWorkspaceVariant } from "@/lib/domain/workspaces/sidebar/sidebar-indicators";
 
 const SIDEBAR_WORKSPACE_TYPE_OPTIONS: Array<{
