+++ apps/packages/product-domain/src/sessions/activity.ts
@@ -1,3 +1,24 @@
-export * from "./activity-types";
-export * from "./activity-status";
-export * from "./activity-workspace";
+export type {
+  SessionViewState,
+  SidebarSessionActivityState,
+} from "./activity-types";
+export {
+  isSessionEffectivelyStreaming,
+  isSessionSlotBusy,
+  resolveSessionErrorAttentionKey,
+  resolveSessionExecutionPhase,
+  resolveSessionSidebarActivityState,
+  resolveSessionStatus,
+  resolveSessionViewState,
+  resolveStatusFromExecutionSummary,
+  resolveWorkspaceExecutionSidebarActivityState,
+  resolveWorkspaceExecutionViewState,
+  shouldSkipColdIdleSessionStream,
+} from "./activity-status";
+export {
+  collectSessionActivityReconciliationIds,
+  collectWorkspaceSessionViewStates,
+  collectWorkspaceSidebarActivityStates,
+  collectWorkspaceSidebarActivityStatesWithErrorAttention,
+  sessionSlotBelongsToWorkspace,
+} from "./activity-workspace";
