+++ apps/packages/product-domain/src/workspaces/cloud-work-inventory.ts
@@ -1,8 +1,67 @@
-export * from "./cloud-work-inventory-types";
-export * from "./cloud-work-filters";
-export * from "./cloud-work-items";
-export * from "./cloud-work-labels";
-export * from "./cloud-work-runtime";
-export * from "./cloud-work-status";
-export * from "./cloud-work-time";
-export * from "./recent-work-items";
+export type {
+  BuildCloudWorkInventoryOptions,
+  BuildRecentWorkItemsOptions,
+  CloudCommandReadinessState,
+  CloudCommandReadinessView,
+  CloudWorkFilters,
+  CloudWorkGroupView,
+  CloudWorkItemView,
+  CloudWorkOpenTarget,
+  CloudWorkOwnerFilter,
+  CloudWorkOwnerKind,
+  CloudWorkRecencyGroupId,
+  CloudWorkRecencyGroupView,
+  CloudWorkSort,
+  CloudWorkSource,
+  CloudWorkStatusFilter,
+  CloudWorkspaceLastSessionSummary,
+  RecentWorkCloudAccessState,
+  RecentWorkCommandability,
+  RecentWorkItemView,
+  RecentWorkOpenTarget,
+  RecentWorkOwnership,
+  RecentWorkRowKind,
+  RecentWorkRuntimeLocation,
+  RecentWorkSourceKind,
+  RecentWorkState,
+  RecentWorkStatusIndicatorKind,
+  RecentWorkStatusIndicatorTone,
+  RecentWorkStatusIndicatorView,
+} from "./cloud-work-inventory-types";
+export { CLOUD_WORK_SOURCE_ORDER } from "./cloud-work-inventory-types";
+export {
+  buildCloudWorkInventory,
+  buildCloudWorkRecencyInventory,
+  compareCloudWorkItems,
+  compareCloudWorkItemsForSort,
+  filterCloudWorkItems,
+  groupCloudWorkItemsByRecency,
+} from "./cloud-work-filters";
+export {
+  cloudWorkActivityPreview,
+  cloudWorkItemForWorkspace,
+  cloudWorkSourceAgentKind,
+  cloudWorkSourceForWorkspace,
+  recentWorkSourceForWorkspace,
+} from "./cloud-work-items";
+export {
+  recentWorkCloudAccessLabel,
+  recentWorkCommandabilityLabel,
+  recentWorkRuntimeLabel,
+  recentWorkSourceLabel,
+} from "./cloud-work-labels";
+export {
+  cloudCommandReadiness,
+  cloudWorkspaceRuntimeIsInProgress,
+  recentWorkCloudAccessState,
+  recentWorkCommandability,
+  recentWorkRuntimeLocationForWorkspace,
+} from "./cloud-work-runtime";
+export {
+  cloudWorkStatusForWorkspace,
+  recentWorkStatusIndicatorForSession,
+  recentWorkStatusIndicatorForWorkspace,
+  selectDefaultCloudWorkSession,
+} from "./cloud-work-status";
+export { dedupeCloudWorkspaces } from "./cloud-work-time";
+export { buildRecentWorkItems } from "./recent-work-items";
