+++ apps/packages/product-domain/src/workspaces/inventory.ts
@@ -1,6 +1,34 @@
-export * from "./workspace-inventory-types";
-export * from "./workspace-inventory-options";
-export * from "./workspace-inventory-builders";
-export * from "./workspace-inventory-filters";
-export * from "./workspace-inventory-groups";
-export * from "./workspace-inventory-summary";
+export type {
+  BuildCloudWorkspaceInventoryOptions,
+  WorkspaceInventoryFilterId,
+  WorkspaceInventoryFilterOption,
+  WorkspaceInventoryGroupBy,
+  WorkspaceInventoryGroupOption,
+  WorkspaceInventoryGroupView,
+  WorkspaceInventoryItemView,
+  WorkspaceInventoryLocationKind,
+  WorkspaceInventoryOwnershipKind,
+  WorkspaceInventorySourceKind,
+  WorkspaceInventoryStatusFilterKind,
+  WorkspaceInventoryStatusKind,
+} from "./workspace-inventory-types";
+export { WORKSPACE_INVENTORY_GROUP_OPTIONS } from "./workspace-inventory-options";
+export {
+  buildCloudWorkspaceInventoryGroups,
+  buildCloudWorkspaceInventoryItems,
+} from "./workspace-inventory-builders";
+export {
+  buildWorkspaceInventoryFilterOptions,
+  filterWorkspaceInventoryItems,
+} from "./workspace-inventory-filters";
+export {
+  groupInventoryByOwnership,
+  groupInventoryByRuntime,
+  groupInventoryBySource,
+  groupInventoryByStatus,
+  groupWorkspaceInventoryItems,
+} from "./workspace-inventory-groups";
+export {
+  workspaceInventorySummaryLabel,
+  workspaceInventorySyncLabel,
+} from "./workspace-inventory-summary";
