+++ apps/packages/product-surfaces/package.json
@@ -19,6 +19,10 @@
       "types": "./dist/settings/CloudEnvironmentsSettingsSurface.d.ts",
       "import": "./dist/settings/CloudEnvironmentsSettingsSurface.js"
     },
+    "./settings/cloud-environments/AddCloudEnvironmentDialogController": {
+      "types": "./dist/settings/cloud-environments/AddCloudEnvironmentDialogController.d.ts",
+      "import": "./dist/settings/cloud-environments/AddCloudEnvironmentDialogController.js"
+    },
     "./support/CloudSupportSurface": {
       "types": "./dist/support/CloudSupportSurface.d.ts",
       "import": "./dist/support/CloudSupportSurface.js"
