+++ apps/packages/product-ui/package.json
@@ -191,6 +191,34 @@
       "types": "./dist/sidebar/ProductSidebar.d.ts",
       "import": "./dist/sidebar/ProductSidebar.js"
     },
+    "./sidebar/ProductSidebarFooter": {
+      "types": "./dist/sidebar/ProductSidebarFooter.d.ts",
+      "import": "./dist/sidebar/ProductSidebarFooter.js"
+    },
+    "./sidebar/ProductSidebarLayout": {
+      "types": "./dist/sidebar/ProductSidebarLayout.d.ts",
+      "import": "./dist/sidebar/ProductSidebarLayout.js"
+    },
+    "./sidebar/ProductSidebarModel": {
+      "types": "./dist/sidebar/ProductSidebarModel.d.ts",
+      "import": "./dist/sidebar/ProductSidebarModel.js"
+    },
+    "./sidebar/ProductSidebarNavigation": {
+      "types": "./dist/sidebar/ProductSidebarNavigation.d.ts",
+      "import": "./dist/sidebar/ProductSidebarNavigation.js"
+    },
+    "./sidebar/ProductSidebarRepositories": {
+      "types": "./dist/sidebar/ProductSidebarRepositories.d.ts",
+      "import": "./dist/sidebar/ProductSidebarRepositories.js"
+    },
+    "./sidebar/ProductSidebarShowToggleRow": {
+      "types": "./dist/sidebar/ProductSidebarShowToggleRow.d.ts",
+      "import": "./dist/sidebar/ProductSidebarShowToggleRow.js"
+    },
+    "./sidebar/ProductSidebarThreads": {
+      "types": "./dist/sidebar/ProductSidebarThreads.d.ts",
+      "import": "./dist/sidebar/ProductSidebarThreads.js"
+    },
     "./new-chat/NewChatSurface": {
       "types": "./dist/new-chat/NewChatSurface.d.ts",
       "import": "./dist/new-chat/NewChatSurface.js"
