+++ apps/web/src/components/home/screen/HomeScreen.tsx
@@ -52,7 +52,7 @@ import type {
 import type { CloudChatComposerControlView } from "@proliferate/product-ui/chat/CloudChatComposer";
 import { NewChatSurface } from "@proliferate/product-ui/new-chat/NewChatSurface";
 import type { CloudChatTranscriptRowView } from "@proliferate/product-domain/chats/cloud/transcript-view";
-import { AddCloudEnvironmentDialogController } from "@proliferate/product-surfaces/settings/CloudEnvironmentsSettingsSurface";
+import { AddCloudEnvironmentDialogController } from "@proliferate/product-surfaces/settings/cloud-environments/AddCloudEnvironmentDialogController";
 
 import { webEnv } from "../../../config/env";
 import { routes } from "../../../config/routes";
